requirements.txt
cryptography==39.0.0
python-dotenv==1.0.0
qrcode==7.3.1
streamlit==1.22.0